+++ src/v2/refinement/RefinedString.spec.ts
@@ -31,8 +31,9 @@
 // ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 // POSSIBILITY OF SUCH DAMAGE.
 //
-import { OnError } from "@ganbarodigital/ts-on-error/V1";
+import { AnyAppError, OnError } from "@ganbarodigital/ts-lib-error-reporting/lib/v1";
 
+import { NeverABrandedUuidError } from "../fixtures";
 import { ValueObject } from "../types/ValueObject";
 import { RefinedString } from "./RefinedString";
 
@@ -41,10 +42,10 @@ function mustBeUuid(input: string): void {
 }
 
 function neverAUuid(input: string, onError: OnError): never {
-    throw onError(Symbol("neverAUuid"), "never a UUID", {input});
+    throw onError(new NeverABrandedUuidError());
 }
 
-function defaultOnErrorHandler(reason: symbol, description: string, extra: object): never {
+function defaultOnErrorHandler(e: AnyAppError): never {
     throw new Error("DEFAULT ERROR HANDLER CALLED");
 }